requirements.txt
Flask ~= 0.10
PyGithub ~= 1.26
